Added Deck class and changed Card class
Deck class contains private linkedlist holding the cards in the deck. Added methods to shuffle the deck, add or remove a card to the top or bottom of the deck, and getter for the size of the linkedlist. Switched the suit of the Card class from integer keys to an enumeration. Made minor changes to CardTest.java