Added splitDeck() method to Deck class & tests
splitDeck() removes bottom half of deck from the deck, and returns the bottom half. If the amount of cards in the deck is odd then the top/first half will be the larger of the two.
Added DeckTest.java to test the splitDeck method with both even and odd decks.
Added takeFrom(int) method to remove and return a given card from a Deck.
Added cardAt(int) method to reference cards LinkedList within each Deck. (same as deck.cards.get(int) if cards were visible).
Added a toString method to Card to make debugging easier.