Created abstract Game class, Display class, and MouseInputHandler class
Game will serve as a template for each type of game which will inherit Game.
It's serializable, has abstract methods drawGraphics(Graphics2D) and handleMouseInput(MouseEvent), and a saveGame(String) method to save the game to a given file. It has two transient members, references to its Display and MouseInputHandler.

Display extends JPanel and will be used to paint the game onto.
It has one private member, game, a reference to the Game that paints on it. It overrides the paint method of JPanel and calls the drawGraphics(Graphics2D) method of game with the graphics context after casting to Graphics2D and enabling antialiasing.

MouseInputHandler implements MouseListener and MouseMotionListener.
It has one private member, game, a reference to the Game that the input is handled by. Every overridden method calls the handleMouseInput(MouseEvent) of game.